internal/cmd: split builtin metadata generation from capabilities
It's cleaner. Plus, if we add examples, that might happen there, too,
and it's quite unrelated to capabilities.
